Ajustando get de plano ciclo

+++ src/SME.SGP.Dto/PlanoCicloDto.cs
@@ -6,7 +6,7 @@ namespace SME.SGP.Dto
     public class PlanoCicloDto
     {
         [Required(ErrorMessage = "O ano deve ser informado")]
-        public long Ano { get; set; }
+        public int Ano { get; set; }
 
         [Required(ErrorMessage = "O ciclo deve ser informado")]
         public long CicloId { get; set; }
